implement board like

+++ CarTalkProject/src/main/java/com/kh/spring/board/model/dto/LikeDTO.java
@@ -13,7 +13,7 @@ import lombok.ToString;
 @ToString
 public class LikeDTO {
 
-	private Long userNo;
+	private int userNo;
 	private Long boardNo;
 	
 }
